+++ sbi/neural_nets/factory.py
@@ -28,15 +28,12 @@ from sbi.neural_nets.net_builders.flow import (
     build_zuko_unconditional_flow,
 )
 from sbi.neural_nets.net_builders.mdn import build_mdn
-from sbi.neural_nets.net_builders.mixed_nets import build_mnle, build_mnpe
+from sbi.neural_nets.net_builders.mnle import build_mnle
 from sbi.neural_nets.net_builders.vector_field_nets import (
-    build_mlp_vector_field,
-    build_transformer_cross_attn_vector_field,
-    build_transformer_vector_field,
+    build_mlp_flowmatcher,
+    build_resnet_flowmatcher,
+    build_score_estimator,
 )
-from sbi.neural_nets.net_builders.mdn import build_mdn
-from sbi.neural_nets.net_builders.mixed_nets import build_mnle, build_mnpe
-from sbi.neural_nets.net_builders.score_nets import build_score_estimator
 from sbi.utils.nn_utils import check_net_device
 
 model_builders = {
